requirements.txt
# Core dependencies
Pillow==10.2.0              # For image processing and steganography
colorama==0.4.6             # Colored CLI output (optional aesthetic)
pytest==8.3.1               # Unit testing framework

# Optional extended cryptography support (for future)
cryptography==43.0.1

# Documentation utilities (optional)
markdown==3.6
